+++ src/compas_viewer/ui/__init__.py
@@ -1,3 +1,0 @@
-from .ui import UI
-
-__all__ = ["UI"]